SA - 3.1.0.49 - 4/4/2013 (stereo branch) 1) Fixed problem in autopitch graph where graph was double the length from what it should be. 2) Fixed crash in spectrogram graph when 96khz file is opened -  Code was truncating data value. 3) Fixed crash in autopitch graph when 96khz file was opened - same as #2. Known Problems: 1) Doing a cut/paste as new/undo corrupts the original graph. 2) The spectrogram graph is corrupted when processing 96khz files. 3) The Formant graphs don't work.

+++ Src/SaExe/DSP/Signal.h
@@ -18,18 +18,18 @@ typedef struct SIG_PARMS
     void * Start;
     uint32 Length;
     int8 SmpDataFmt;
-    uint16 SmpRate;
+    uint32 SmpRate;
     SIG_SOURCE Source;
 } SIG_PARMS;
 
 // PROC_PARMS added by CLW 1/5/99
 typedef struct
 {
-    void    *    pBufferStart;  // pointer to buffer containing data to be processed
-    uint32    dwBufferSize;  // size of buffer in bytes
-    uint32    dwBlockOffset;  // sample number at which processing begins
-    uint32    dwBlockLength;  // number of samples in process data
-    SIG_PARMS  stSigParms;
+    void * pBufferStart;	// pointer to buffer containing data to be processed
+    uint32 dwBufferSize;	// size of buffer in bytes
+    uint32 dwBlockOffset;	// sample number at which processing begins
+    uint32 dwBlockLength;	// number of samples in process data
+    SIG_PARMS stSigParms;
 } PROC_PARMS;
 
 #endif
